fix(dashboard-menu): close all panels when app is switched
Currently, when we expand the interactions panel and switch to a different
app, the interactions from the previous app remain open. Our existing logic
attempts to redraw the interactions list when the app is switched, but it
doesn't function as expected, especially when dealing with a higher count
of interactions.

From now on, whenever we switch between apps, all currently open panels will
be automatically closed. The manual toggling of panels will ensure accurate
interaction list is being displayed

Jira Issue: DS-473